+++ lib/discordrb/data/member.rb
@@ -6,6 +6,9 @@ module Discordrb
     # @return [Time] when this member joined the server.
     attr_reader :joined_at
 
+    # @return [Time, nil] when this member boosted this server, `nil` if they haven't.
+    attr_reader :boosting_since
+
     # @return [String, nil] the nickname this member has, or `nil` if it has none.
     attr_reader :nick
     alias_method :nickname, :nick
@@ -69,6 +72,12 @@ module Discordrb
 
       @nick = data['nick']
       @joined_at = data['joined_at'] ? Time.parse(data['joined_at']) : nil
+      @boosting_since = data['premium_since'] ? Time.parse(data['premium_since']) : nil
+    end
+
+    # @return [true, false] if this user is a Nitro Booster of this server.
+    def boosting?
+      !@boosting_since.nil?
     end
 
     # @return [true, false] whether this member is the server owner.
@@ -240,6 +249,13 @@ module Discordrb
     # @!visibility private
     def update_nick(nick)
       @nick = nick
+    end
+
+    # Update this member's boosting timestamp
+    # @note For internal user only.
+    # @!visibility private
+    def update_boosting_since(time)
+      @boosting_since = time
     end
 
     include PermissionCalculator