+++ lib/discordrb/data/server.rb
@@ -56,6 +56,14 @@ module Discordrb
     # @return [Hash<Integer => VoiceState>] the hash (user ID => voice state) of voice states of members on this server
     attr_reader :voice_states
 
+    # The server's amount of Nitro boosters.
+    # @return [Integer] the amount of boosters, 0 if no one has boosted.
+    attr_reader :booster_count
+
+    # The server's Nitro boost level.
+    # @return [Integer] the boost level, 0 if no level.
+    attr_reader :boost_level
+
     # @!visibility private
     def initialize(data, bot, exists = true)
       @bot = bot
@@ -68,6 +76,7 @@ module Discordrb
       @large = data['large']
       @member_count = data['member_count']
       @splash_id = nil
+      @banner_id = nil
       @features = data['features'].map { |element| element.downcase.to_sym }
       @members = {}
       @voice_states = {}
@@ -85,6 +94,9 @@ module Discordrb
 
       # Only get the owner of the server actually exists (i.e. not for ServerDeleteEvent)
       @owner = member(@owner_id) if exists
+
+      @booster_count = data['premium_subscription_count'] || 0
+      @boost_level = data['premium_tier']
     end
 
     # The default channel is the text channel on this server with the highest position
@@ -364,6 +376,20 @@ module Discordrb
       API.splash_url(@id, @splash_id)
     end
 
+    # @return [String] the hexadecimal ID used to identify this server's banner image, shown by the server name.
+    def banner_id
+      @banner_id ||= JSON.parse(API::Server.resolve(@bot.token, @id))['banner']
+    end
+
+    # @return [String, nil] the banner image URL for the server's banner image, or
+    #   `nil` if there is no banner image.
+    def banner_url
+      banner_id if @banner_id.nil?
+      return unless banner_id
+
+      API.banner_url(@id, @banner_id)
+    end
+
     # Adds a role to the role cache
     # @note For internal use only
     # @!visibility private
@@ -547,6 +573,21 @@ module Discordrb
       data = JSON.parse(API::Server.edit_emoji(@bot.token, @id, emoji.resolve_id, name || emoji.name, (roles || emoji.roles).map(&:resolve_id), reason))
       new_emoji = Emoji.new(data)
       @emoji[new_emoji.id] = new_emoji
+    end
+
+    # The amount of emoji the server can have, based on its current Nitro Boost Level.
+    # @return [Integer] the max amount of emoji
+    def max_emoji
+      case @level
+      when 1
+        100
+      when 2
+        150
+      when 3
+        250
+      else
+        50
+      end
     end
 
     # @return [Array<ServerBan>] a list of banned users on this server and the reason they were banned.